expose QuickPickService on the backend in the connection scope
